Split menu for each version RS232 Spectranet
