+++ src/snapCterm_RS232.h
@@ -5,4 +5,7 @@ void CommsInit(void);
 void RX(void);
 void TX(void);
 
+void Draw_Menu(void);
+void menu(void);
+
 #endif
