+++ src/snapCterm_SNet.h
@@ -11,4 +11,7 @@ void CommsInit(void);
 void RX(void);
 void TX(void);
 
+void Draw_Menu(void);
+void menu(void);
+
 #endif
